Lockscreen charging info: show decimal in battery temperature

+++ packages/SettingsLib/src/com/android/settingslib/fuelgauge/BatteryStatus.java
@@ -58,13 +58,13 @@ public class BatteryStatus {
     public final int maxChargingWattage;
     public final boolean present;
 
-    public final int temperature;
+    public final float temperature;
     public final boolean dashChargeStatus;
 
     public BatteryStatus(int status, int level, int plugged, int health,
             int maxChargingWattage, boolean present,
             int maxChargingCurrent, int maxChargingVoltage,
-            int temperature, boolean dashChargeStatus) {
+            float temperature, boolean dashChargeStatus) {
         this.status = status;
         this.level = level;
         this.plugged = plugged;
